lib: Always fetch data from Sunday Fix uneven graph If I want the data for the last 5 weeks, i.e. today is 'Thursday, 5th Oct 2023' and 5 weeks ago was 'Thursday, 31st August 2023'. Then in this case, fetch data from the Sunday of that week, i.e. 'Sunday, 27th August 2023'.
Why?

What happened here is that in the first column, it would only show 31st Aug, 1st Sep, 2nd Sep. From the 2nd column it would start from Sunday from the top of the column. But for the first column, even if was some day in mid week, it would disregard it and start from the top.

Maybe I should have used Grid here.